server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["Sambhav Kothari <sambhavs.email@gmail.com>"]

[tool.poetry.dependencies]
python = "~3.7"
fastapi = "^0.58.1"
uvicorn = "^0.11.5"
PyGithub = "^1.51"
black = "^19.10b0"
isort = "^5.0.9"
cookiecutter = "^1.7.2"

[tool.poetry.dev-dependencies]
pytest = "^4.6"
mypy = "^0.782"
pudb = "^2019.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
